Changing nugget mining system
Within the execute function of the EnterMineAndDigForNugget state, I removed the actions that
added gold to the miner's gold carried, thirst and fatigue.

The random module is used to introduce some randomness to the results of
miningl.  On each mining change if r < 0.25 the miner will get 4 nuggets,
if r < 0.50 the miner will get 2 nuggets and if r < 0.95 the miner will get 1 nugget.  Otherwise the miner
will get no nuggets. No matter what the r value is the miner will always gain 1 fatigue and 1 thrist.