Added Jail state and minor changes
To allow the miner to have two different exits from the Quench Thirst state I added two attributes to the miner state.  Status toggles between 'free' and 'arrested'.  When checked during the exit function of Quench Thirst, miner's status triggers the proper dialogue.
The miner attribute counter_jail counts the number of game ticks the miner has spent in jail.  The Jail execute function currently changes the miner's status after 3 game ticks.
The thirsty function was changed to return True if greater than 10.  This is because the lower thresh hold was sending him to the saloon too often.
The game_object other_miner was removed from the list to make it simpler to track the progress when the code was run.
Fixed typo in Miner dialogue in Deposit gold execute function.
In Deposit gold execute function, made the thresh hold higher for becoming rich and changing state to go home.  Before, miner would have always entered this if statement because the condition for him to go to the bank in the first place was higher than this number.
Fixed typo in Quench thirst enter function
Quench thirst execute function now has a conditional statement that will send the miner to jail if his gold_carried == 0.  The barkeep will call the sheriff, then the miner's status is changed to 'arrested' and the change_state function is called.
Quench thirst exit function now has a check for miner's status.  If 'arrested', the proper dialogue will execute.
The Jail enter function changes the location to 'jail', and runs a line of dialogue from the miner.
The Jail execute function resets the miner's fatigue and thirst to 0, so that he will start fresh at the mine when he has completed the Jail state.   The miner has a line of dialogue.  The counter_jail attribute is advanced by one and the execute function runs through a loop 3 times before the miner changes states to enter_mine_and_dig_for_nugget.
The Jail exit function resets the counter_jail attribute to 0 and resets the status attribute to 'free'.
The miner has one line of dialogue.