new items file, shop state, and stats for miner
The new file items contains all information on items that the miner can
use. For now the only items in there is the pickax which can influence
the miner's strength and ability to mine gold. To do this the pickax
itself is a class containing with in it's init a variable for strength,
a variable called luck which is what influences the miner's gold finding
abilities, and an id variable to help keep track of them in case more
are created. The new class GoShopping was created so the miner can
purchase a better pickax if certain conditions are met. These conditions
were a certain amount of gold in his savings and the current strength of
his pickax, however since his savings must've been at least 10 the
amount of gold to head home was increased. Once he has made it to the
shop he will purchase a new pickax based on the same stats. After his
purchase he goes directly back to the mines. To add all this in to the
miner he was given a build and pickax variable during his creation and
init which defines his strength, health, and the pickax he weilds at the
beginning. All of this is explained in greater detail in the README.md
file.